Update TextController.cs
Tying in some of the controls so we can play this baby and get it done. Ghost rooms, the empty room, Cipher's Door and Cipher's room and the grabbing the fire extinguisher now need to be tested.

I found a tutorial for adding a gamepad, so let's give that a try.